+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/object-colors/compare/v2.3.0...HEAD)
 ------------------------------------------------------------------------
+### Fixed
+- add default `NoneType` to `__setattr__`
 
 [2.3.0](https://github.com/jshwi/object-colors/releases/tag/v2.3.0) - 2023-09-24
 ------------------------------------------------------------------------