+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-[Unreleased](https://github.com/jshwi/object-colors/compare/v2.3.0...HEAD)
+[Unreleased](https://github.com/jshwi/object-colors/compare/v2.3.1...HEAD)
+------------------------------------------------------------------------
+
+[2.3.1](https://github.com/jshwi/object-colors/releases/tag/v2.3.1) - 2023-09-24
 ------------------------------------------------------------------------
 ### Fixed
 - add default `NoneType` to `__setattr__`